refactor(api-client,core): Rename root to index
