+++ packages/api-client/src/main/notification/NotificationAPI.ts
@@ -19,7 +19,7 @@
 
 import {AxiosRequestConfig} from 'axios';
 import {HttpClient} from '../http';
-import {Notification, NotificationList} from './index';
+import {Notification, NotificationList} from './';
 
 const NOTIFICATION_SIZE_MAXIMUM = 10000;
