+++ packages/core/src/demo/echo.js
@@ -18,7 +18,7 @@ const logger = logdown('@wireapp/core/demo/echo.js', {
 logger.state.isEnabled = true;
 
 const {Account} = require('@wireapp/core');
-const {PayloadBundleType} = require('@wireapp/core/dist/conversation/root');
+const {PayloadBundleType} = require('@wireapp/core/dist/conversation/');
 const {APIClient} = require('@wireapp/api-client');
 const {ClientType} = require('@wireapp/api-client/dist/commonjs/client/ClientType');
 const {ConnectionStatus} = require('@wireapp/api-client/dist/commonjs/connection/');
