+++ packages/core/src/main/Account.ts
@@ -17,9 +17,9 @@
  *
  */
 
-import {Context, LoginData} from '@wireapp/api-client/dist/commonjs/auth/index';
-import {ClientType, RegisteredClient} from '@wireapp/api-client/dist/commonjs/client/index';
-import {IncomingNotification} from '@wireapp/api-client/dist/commonjs/conversation/index';
+import {Context, LoginData} from '@wireapp/api-client/dist/commonjs/auth/';
+import {ClientType, RegisteredClient} from '@wireapp/api-client/dist/commonjs/client/';
+import {IncomingNotification} from '@wireapp/api-client/dist/commonjs/conversation/';
 import {
   CONVERSATION_EVENT,
   ConversationEvent,
@@ -28,16 +28,24 @@ import {
   IncomingEvent,
   USER_EVENT,
   UserEvent,
-} from '@wireapp/api-client/dist/commonjs/event/index';
-import {StatusCode} from '@wireapp/api-client/dist/commonjs/http/index';
-import {WebSocketClient} from '@wireapp/api-client/dist/commonjs/tcp/index';
+} from '@wireapp/api-client/dist/commonjs/event/';
+import {StatusCode} from '@wireapp/api-client/dist/commonjs/http/';
+import {WebSocketClient} from '@wireapp/api-client/dist/commonjs/tcp/';
 import * as cryptobox from '@wireapp/cryptobox';
 import {GenericMessage} from '@wireapp/protocol-messaging';
-import {RecordNotFoundError} from '@wireapp/store-engine/dist/commonjs/engine/error/index';
+import {RecordNotFoundError} from '@wireapp/store-engine/dist/commonjs/engine/error/';
 import * as Long from 'long';
-import {LoginSanitizer} from './auth/root';
-import {ClientInfo, ClientService} from './client/root';
-import {ConnectionService} from './connection/root';
+import {LoginSanitizer} from './auth/';
+import {ClientInfo, ClientService} from './client/';
+import {ConnectionService} from './connection/';
+import {
+  AssetService,
+  ConversationService,
+  GenericMessageType,
+  PayloadBundleIncoming,
+  PayloadBundleState,
+  PayloadBundleType,
+} from './conversation/';
 import {
   AssetContent,
   ClearedContent,
@@ -49,18 +57,10 @@ import {
   ReactionContent,
   TextContent,
 } from './conversation/content/';
-import {
-  AssetService,
-  ConversationService,
-  GenericMessageType,
-  PayloadBundleIncoming,
-  PayloadBundleState,
-  PayloadBundleType,
-} from './conversation/root';
-import {CryptographyService} from './cryptography/root';
-import {GiphyService} from './giphy/root';
-import {NotificationService} from './notification/root';
-import {SelfService} from './self/root';
+import {CryptographyService} from './cryptography/';
+import {GiphyService} from './giphy/';
+import {NotificationService} from './notification/';
+import {SelfService} from './self/';
 
 import {APIClient} from '@wireapp/api-client';
 import {UserConnectionEvent} from '@wireapp/api-client/dist/commonjs/event';
