+++ packages/core/src/main/auth/index.ts
