+++ packages/core/src/main/client/ClientDatabaseRepository.ts
@@ -17,9 +17,9 @@
  *
  */
 
-import {RegisteredClient} from '@wireapp/api-client/dist/commonjs/client/index';
-import {CRUDEngine} from '@wireapp/store-engine/dist/commonjs/engine/index';
-import {CryptographyService} from '../cryptography/root';
+import {RegisteredClient} from '@wireapp/api-client/dist/commonjs/client/';
+import {CRUDEngine} from '@wireapp/store-engine/dist/commonjs/engine/';
+import {CryptographyService} from '../cryptography/';
 import {MetaClient} from './ClientService';
 
 export enum DatabaseStores {
