+++ packages/core/src/main/conversation/ConversationService.ts
@@ -41,7 +41,7 @@ import {
   PayloadBundleState,
   PayloadBundleType,
   ReactionType,
-} from '../conversation/root';
+} from '../conversation/';
 
 import {
   Article,
@@ -85,8 +85,8 @@ import {
 
 import {TextContentBuilder} from './TextContentBuilder';
 
+import {CryptographyService, EncryptedAsset} from '../cryptography/';
 import * as AssetCryptography from '../cryptography/AssetCryptography.node';
-import {CryptographyService, EncryptedAsset} from '../cryptography/root';
 
 import {APIClient} from '@wireapp/api-client';
