+++ packages/core/src/main/conversation/content/AssetContent.ts
@@ -17,9 +17,9 @@
  *
  */
 
+import {AbortReason, AssetTransferState} from '../../conversation/';
 import {FileContent, FileMetaDataContent, ImageContent} from '../../conversation/content/';
-import {AbortReason, AssetTransferState} from '../../conversation/root';
-import {EncryptedAssetUploaded} from '../../cryptography/root';
+import {EncryptedAssetUploaded} from '../../cryptography/';
 
 // https://github.com/wireapp/generic-message-proto/blob/v1.20.0/proto/messages.proto#L201
 interface AssetContent {
