+++ packages/core/src/main/cryptography/AssetCryptography.browser.ts
@@ -17,7 +17,7 @@
  *
  */
 
-import {EncryptedAsset} from '../cryptography/root';
+import {EncryptedAsset} from '../cryptography/';
 const {crypto} = window;
 
 const isEqual = (a: Buffer, b: Buffer): boolean => {
