+++ packages/core/src/main/cryptography/AssetCryptography.node.ts
@@ -18,7 +18,7 @@
  */
 
 import * as crypto from 'crypto';
-import {EncryptedAsset} from '../cryptography/root';
+import {EncryptedAsset} from '../cryptography/';
 
 const isEqual = (a: Buffer, b: Buffer): boolean => {
   const arrayA = new Uint32Array(a);
