+++ packages/core/src/main/cryptography/CryptographyService.test.node.js
@@ -24,7 +24,7 @@ const bazinga64 = require('bazinga64');
 const crypto = require('crypto');
 const Proteus = require('@wireapp/proteus');
 const {Cryptobox} = require('@wireapp/cryptobox');
-const {CryptographyService} = require('@wireapp/core/dist/cryptography/root');
+const {CryptographyService} = require('@wireapp/core/dist/cryptography/');
 const {MemoryEngine} = require('@wireapp/store-engine');
 const {promisify} = require('util');
 const {decryptAsset, encryptAsset} = require('@wireapp/core/dist/cryptography/AssetCryptography.node');
