+++ packages/core/src/main/index.ts
@@ -18,9 +18,9 @@
  */
 
 import {Account} from './Account';
-import * as auth from './auth/root';
-import * as conversation from './conversation/root';
-import * as cryptography from './cryptography/root';
+import * as auth from './auth/';
+import * as conversation from './conversation/';
+import * as cryptography from './cryptography/';
 
 export = {
   Account,
