+++ packages/core/src/main/notification/NotificationDatabaseRepository.ts
@@ -17,8 +17,8 @@
  *
  */
 
-import {Notification, NotificationEvent} from '@wireapp/api-client/dist/commonjs/notification/index';
-import {CRUDEngine} from '@wireapp/store-engine/dist/commonjs/engine/index';
+import {Notification, NotificationEvent} from '@wireapp/api-client/dist/commonjs/notification/';
+import {CRUDEngine} from '@wireapp/store-engine/dist/commonjs/engine/';
 import CryptographyDatabaseRepository from '../cryptography/CryptographyDatabaseRepository';
 
 export enum DatabaseStores {
