+++ packages/core/src/main/notification/NotificationService.test.browser.js
@@ -18,7 +18,7 @@
  */
 
 const UUID = require('pure-uuid');
-const {NotificationService} = require('@wireapp/core/dist/notification/root');
+const {NotificationService} = require('@wireapp/core/dist/notification/');
 const {IndexedDBEngine} = require('@wireapp/store-engine');
 const {APIClient} = require('@wireapp/api-client');
