+++ packages/cryptobox/src/main/error/index.ts
