+++ packages/cryptobox/src/main/index.ts
@@ -1,9 +1,9 @@
 import Cryptobox from './Cryptobox';
 import CryptoboxSession from './CryptoboxSession';
 import DecryptionError from './DecryptionError';
-import * as error from './error/root';
+import * as error from './error/';
 import InvalidPreKeyFormatError from './InvalidPreKeyFormatError';
 import {SerializedCryptobox} from './SerializedCryptobox';
-import * as store from './store/root';
+import * as store from './store/';
 
 export {Cryptobox, CryptoboxSession, DecryptionError, InvalidPreKeyFormatError, error, store, SerializedCryptobox};
