+++ packages/proteus/src/main/errors/index.ts
